refactor package style
