+++ mono/src/main/java/com/wanted/mono/domain/product/entity/ProductStatus.java
@@ -1,4 +1,4 @@
-package com.wanted.mono.model.entity;
+package com.wanted.mono.domain.product.entity;
 
 public enum ProductStatus {
     판매중, 품절, 삭제됨
