+++ mono/src/main/java/com/wanted/mono/domain/user/entity/User.java
@@ -1,36 +1,32 @@
-package com.wanted.mono.model.entity;
+package com.wanted.mono.domain.user.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.hibernate.annotations.CreationTimestamp;
+
+import java.time.LocalDateTime;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-@Table(name = "brands")
-public class Brand {
-    // 브랜드 ID (PK)
+@Table(name = "users")
+public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    // 브랜드명
     @Column(length = 100, nullable = false)
     private String name;
 
-    // URL 슬러그
     @Column(length = 100, unique = true, nullable = false)
-    private String slug;
+    private String email;
 
-    // 설명
-    private String description;
-
-    // 로고 이미지 URL
     @Column(length = 255)
-    private String logoUrl;
+    private String avatarUrl;
 
-    // 웹사이트 URL
-    @Column(length = 255)
-    private String website;
+    @CreationTimestamp
+    @Column(columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP")
+    private LocalDateTime createdAt;
 }
